requirements.txt
pandas
numpy
scikit-learn
pyyaml
google-cloud-storage
joblib
mlflow
fastapi
uvicorn[standard]

